backend/requirements.txt
Flask
Flask-Cors
pymongo
yara-python
typing
uuid
datetime
